requirements.txt
requests~=2.26.0

Flask~=2.0.1
websocket-client~=0.58.0
websockets~=9.1
setuptools~=57.4.0
pandas~=1.3.1